requirements.txt
aiohttp~=3.11
aiofiles~=24.1
requests~=2.32
pydub~=0.25
jiwer~=3.1
httpx~=0.28
pytest~=8.3
pytest-asyncio~=0.25
deepgram-sdk~=3.10
python-dotenv~=1.0
urllib3~=1.25.3, < 3.0.0
python_dateutil~=2.8.2
pydantic~=2
typing-extensions~=4.7.1
websocket-client~=1.8.0